Push Notifications (#69)
# Push Notifications and Account Settings

## :recycle: Current situation & Problem
Right now, we only support in-app notifications. Ideally, we would like
to register the device for remote notifications from the server using
Apple Push Notification service (APNS) as part of the onboarding
process, then send push notifications with a simple action associated
with them (just like the in-app messages). We also need to fetch the
contact information for the organization associated with the current
user and display it in the Account sheet, as well as provide the option
to disable notifications in the Account sheet.


## :gear: Release Notes 
- Implemented a `UserMetaDataManager` module that registers a snapshot
listener to the current user's document in the Users collection in
Firestore
- The snapshot listener decodes the user's notification settings (see
https://github.com/StanfordBDHG/ENGAGE-HF-Firebase/pull/90 for the data
model), as well as the contact information for the organization
associated with the user.
- Added `Contacts` and `Notifications` rows to Account sheet that link
to views that let the user call/email the site leader for their
organization and toggle whether or not to receive each notification
type, respectively. When no contact info is found, the contacts page
shows a basic `ContentUnavailableView`.
- Implemented a `NotificationManager` that conforms to Spezi's
`NotificationHandler` and manages remote notification registration (via
a call to a cloud function that sends the device token to the server),
notification actions (triggered when the user taps on the notification),
and updates to the device token. Notification authorization is requested
in the `NotificationPermission` view, as part of the onboarding flow,
but currently disabled for the simulator.
- Notification actions are the same strings as `MessageAction`.


## :books: Documentation
See inline documentation.


## :white_check_mark: Testing
Includes UI Testing for the contacts view, but it is not able to cover
the decoding portion because the CI does not yet use the seeded
emulator. However, this is not a problem for notification settings,
since the notification setting decoding allows for optional fields, so
UI Testing covers the notification settings view and verifies that
toggling works as expected.

For push notifications, due to the fact that registering for APNS
requires signing from an Apple Developer account, we are unable to test
registering for notifications or receiving remote notifications from the
server until the app is built in a production environment on a mobile
device. However, we were able to confirm that notification actions work
as expected using the "drag and drop" method built into the simulator
with a .apns file.


### Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md):
- [X] I agree to follow the [Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md).