Fix/notification alerts (#77)
# Surface notification alerts in onboarding flow.

## :recycle: Current situation & Problem
Currently, we continue to the next step of the onboarding flow after
selecting "Allow" in `NotificationPermissions` regardless of whether
there was an error or not. As a result, when there is an error, the
alert briefly appears and then is dismissed, which is not ideal
behavior.

Also, we currently only have a mechanism for registering a device for
notifications. We never "unregister" the device. This is an issue in the
case where a user signs in to the app, registering the device for
notifications connected to their account, then signs out. When the next
user signs in, the device is still registered to the account of the
previous user. Instead, we should unregister the device on sign out to
prevent multiple accounts' worth of notifications being associated with
a device.


## :gear: Release Notes 
- Only continue to next step after successfully registering for remote
notifications.
- Added a skip button as a secondary action that just moves on to the
next onboarding step without asking for notification permissions.
- Added logic to request notification permissions on sign in and
unregister on sign out via cloud function calls.
- Also removed Consent sets from onboarding flow, as at the moment.


## :books: Documentation
See inline documentation.


## :white_check_mark: Testing
The changes have been manually tested in addition to the testing already
in place.


### Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md):
- [X] I agree to follow the [Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md).