Heart Health UI Tweaks (#80)
# Make Graph Point Selection More Forgiving, Handle Empty Content

## :recycle: Current situation & Problem
Currently, the graph interactions find a point that was tapped on by
looking for a point that lies in the same time interval as the tap
location (as determined by the selected date resolution). For example,
if the date resolution is monthly, and I tap on a date in August, the
point that is selected would be a point that lies in August.

However, this means that in situations where the interval is small
relative to the size of the date domain of the chart, it can be
difficult to select points because the target interval spans such a tiny
area of the screen.


## :gear: Release Notes 
- Modified the point selection function to look for the point nearest to
the start of the interval that was tapped on, and selecting that date if
it is within a pre-defined margin. This allows the point selection to be
more forgiving.
- The margin is defined as 4% of the size of the date domain of the
chart (this allows the forgiveness to be consistent across different
domain sizes).
- Instead of displaying text that says something along the lines of "No
data, please take a measurement" when the vitals graphs have no data to
display, we now show an empty chart with "No Data" in the header.
- Similarly for the All Data list section below the graphs, we now show
an `ContentUnavailableView` with a short description asking the user to
connect a device and begin taking measurements instead of just raw text
when there are no measurements yet.
- Fixed a bug where the y-axis range in Symptom Scores is locked to
0...100 (even for Dizziness which only ranges from 0-5).


## :books: Documentation
See inline documentation.


## :white_check_mark: Testing
The changes are covered where possible by the existing testing framework
(e.g. ensuring that the empty content views appear where appropriate).


### Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md):
- [X] I agree to follow the [Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md).