Independent Enrollment (#144)
# Independent Enrollment

## :recycle: Current situation & Problem
As part of the independent enrollment functionality, we now allow
updating more information when a user is flagged as selfManaged. This
updating of data is controlled by the Firebase functions in that they
provide a questionnaire to update the data and only the id is known to
the iOS app. The questionnaire result is then stored in Firestore as
usual and processed by the functions with a database trigger again.


## :gear: Release Notes
*Add a bullet point list summary of the feature and possible migration
guides if this is a breaking change so this section can be added to the
release notes.*
*Include code snippets that provide examples of the feature implemented
or links to the documentation if it appends or changes the public
interface.*


## :books: Documentation
*Please ensure that you properly document any additions in conformance
to [Spezi Documentation
Guide](https://github.com/StanfordSpezi/.github/blob/main/DOCUMENTATIONGUIDE.md).*
*You can use this section to describe your solution, but we encourage
contributors to document your reasoning and changes using in-line
documentation.*


## :white_check_mark: Testing
- Manual testing possible by running the Firebase emulators from
[ENGAGE-HF-Firebase](https://github.com/StanfordBDHG/ENGAGE-HF-Firebase)
on branch `independent-enrollment` with account
`engagehf-patient2@stanford.edu` (which is already marked as
self-managed)



https://github.com/user-attachments/assets/baf40892-0224-4419-93dc-699133cb9b8e

### Code of Conduct & Contributing Guidelines
By creating and submitting this pull request, you agree to follow our
[Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md).